fix(wen): remove requireModule, requireModuleAsync of mts
