+++ .changeset/twenty-singers-talk.md
@@ -1,0 +1,5 @@
+---
+"@lynx-js/web-mainthread-apis": minor
+---
+
+fix: mts lynx will no longer provide requireModule, requireModuleAsync methods, which is aligned with Native.
