+++ packages/web-platform/web-mainthread-apis/src/MainThreadRuntime.ts
@@ -18,6 +18,7 @@ import {
   type RpcCallType,
   type postExposureEndpoint,
   type LynxContextEventTarget,
+  type LynxJSModule,
 } from '@lynx-js/web-constants';
 import { globalMuteableVars } from '@lynx-js/web-constants';
 import { createMainThreadLynx, type MainThreadLynx } from './MainThreadLynx.js';
@@ -106,7 +107,7 @@ export class MainThreadRuntime {
     public config: MainThreadConfig,
   ) {
     this.__globalProps = config.globalProps;
-    this.lynx = createMainThreadLynx(config, this);
+    this.lynx = createMainThreadLynx(config);
     /**
      * now create the style content
      * 1. flatten the styleInfo
@@ -220,7 +221,19 @@ export class MainThreadRuntime {
 
   __LoadLepusChunk: (path: string) => boolean = (path) => {
     try {
-      this.lynx.requireModule(path);
+      // @ts-expect-error
+      if (self.WorkerGlobalScope) {
+        const lepusChunkUrl = this.config.lepusCode[`${path}`];
+        if (lepusChunkUrl) path = lepusChunkUrl;
+        // @ts-expect-error
+        importScripts(path);
+        const entry = (globalThis.module as LynxJSModule).exports;
+        entry?.(this);
+      } else {
+        throw new Error(
+          'importing scripts synchronously is only available for the multi-thread running mode',
+        );
+      }
       return true;
     } catch {
     }
